requirements.txt
# Root requirements for deployment (Streamlit Cloud / local venv)
streamlit>=1.28.0
pandas>=2.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
numpy>=1.21.0
plotly>=5.0.0
sqlalchemy>=2.0.0
typing-extensions>=4.0.0
dataclasses-json>=0.5.0
toml>=0.10.0

# PostgreSQL support for Google Cloud
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-cloud-secret-manager>=2.16.0
